- major redesign of exporting results.  Now, will use a TriggerToExport value that triggers the Exporter to dump its results either to SQLite or Textfile. - Added Sqlite functionality  (using a .NET sqlite wrapper library) - Added new IsosResultExporters - Added new ScanResultExporters - Improved PeakListExporters

+++ DeconTools.Backend/Data/UIMFSQLiteIsosExporter.cs
@@ -109,7 +109,7 @@ namespace DeconTools.Backend.Data
                 fp.frame_num = (ushort)uimfResult.FrameSet.PrimaryFrame;
                 fp.ims_scan_num = (ushort)getScanNumber(uimfResult.ScanSet.PrimaryScanNumber);
                 fp.charge = (byte)uimfResult.IsotopicProfile.ChargeState;
-                fp.abundance = (ushort)uimfResult.IsotopicProfile.GetAbundance();
+                fp.abundance = (uint)uimfResult.IsotopicProfile.GetAbundance();
                 fp.mz = uimfResult.IsotopicProfile.GetMZ();
                 fp.fit = (float)uimfResult.IsotopicProfile.GetScore();
                 fp.average_mw = uimfResult.IsotopicProfile.AverageMass;
